Date field verification now works correctly.
It requires further testing because I can't control the current time. If anyone wants to mess around with it, the cases that need testing are setting it to 12:00 AM or PM when the current time is 12:00 AM or 12:00 PM.
Just try to break it and see it you get the expected response of either pass or fail.