Start game button is no longer initially disabled in an edit game page (because the user should be able to start a valid game without making any changes). It still starts out disabled in a create game page because the fields start out blank.
Also added an error message if the user tries to click the disabled save button when they first open an edit game page that says "No changes have been made".
This is because they shouldn't have the option to save changes that weren't made (and obviously the button will become enabled once a change is made).