Implements sending requirements to the requirement manager.
Requirements we create now get added to the requirement manager when we submit a final estimate for those requirements. The import requirements table will now also only add requirements that do not have estimates that aren't zero. If you try to create a requirement with the exact same name and description of one that already exists in the requirement manager, it doesn't let it happen and prints an error message. (the error message doesn't show up as of this commit)